code cleanup
had to fix some classes not following contribution guidlines aswell
