+++ src/main/java/net/apolloclient/events/impl/player/AttackEntityEvent.java
@@ -1,14 +1,18 @@
 package net.apolloclient.events.impl.player;
 
+import lombok.Getter;
 import net.apolloclient.events.Event;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.EntityPlayer;
 
-public class AttackEntityEvent extends Event
-{
-    public final Entity target;
-    public AttackEntityEvent(EntityPlayer player, Entity target)
-    {
+/** Fired when player hits entity.
+ * @author MatthewTGM | MatthewTGM#4058
+ * @since 1.0.0 **/
+public class AttackEntityEvent extends Event {
+
+    @Getter private final Entity target;
+
+    /** @param target entity hit. */
+    public AttackEntityEvent(Entity target) {
         this.target = target;
     }
 }
