+++ src/main/java/net/apolloclient/modules/Category.java
@@ -23,5 +23,5 @@ package net.apolloclient.modules;
  * @author Icovid | Icovid#3888
  * @since 1.0.0 **/
 public enum Category {
-    UTIL
+    RENDER, UTIL, PLAYER, GAMEPLAY;
 }
