--- src/main/java/net/apolloclient/modules/impl/util/WindowedFullscreenModule.java
@@ -1,65 +1,0 @@
-/*⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤
- Copyright (C) 2020-2021 developed by Icovid and Apollo Development Team
-
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU Affero General Public License as published
- by the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
-
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU Affero General Public License for more details.
-
- You should have received a copy of the GNU Affero General Public License
- along with this program.  If not, see https://www.gnu.org/licenses/.
-
- Contact: Icovid#3888 @ https://discord.com
- ⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤⏤*/
-
-package net.apolloclient.modules.impl.util;
-
-import net.apolloclient.events.bus.EventSubscriber;
-import net.apolloclient.events.impl.client.GameLoopEvent;
-import net.apolloclient.modules.Category;
-import net.apolloclient.modules.Module;
-import net.minecraft.client.Minecraft;
-import org.lwjgl.LWJGLException;
-import org.lwjgl.opengl.Display;
-import org.lwjgl.opengl.DisplayMode;
-
-
-public class WindowedFullscreenModule extends Module {
-    private boolean lastFullscreen;
-
-    public WindowedFullscreenModule() {
-        super("WindowedFullscreen", "Lets you use multiple monitors without going out of the Fullscreen", Category.UTIL, true);
-    }
-
-    @EventSubscriber
-    public void onTick(GameLoopEvent event) {
-        boolean fullScreenNow = Minecraft.getMinecraft().isFullScreen();
-        if (this.lastFullscreen != fullScreenNow) {
-            this.fullScreen(fullScreenNow);
-            this.lastFullscreen = fullScreenNow;
-        }
-    }
-
-    public void fullScreen(boolean fullscreen) {
-        try {
-            if (fullscreen) {
-                System.setProperty("org.lwjgl.opengl.Window.undecorated", "true");
-                Display.setDisplayMode(Display.getDesktopDisplayMode());
-                Display.setLocation(0, 0);
-                Display.setFullscreen(false);
-                Display.setResizable(false);
-            } else {
-                System.setProperty("org.lwjgl.opengl.Window.undecorated", "false");
-                Display.setDisplayMode(new DisplayMode(Minecraft.getMinecraft().displayWidth, Minecraft.getMinecraft().displayHeight));
-                Display.setResizable(true);
-            }
-        } catch (LWJGLException lwjglexception) {
-            lwjglexception.printStackTrace();
-        }
-    }
-}
